Add files via upload
Added functionality that stores the winner of the last hand, changes offset from dealer to last winner -1.  This means that instead of the start of a trick going to the person one over from the dealer (dealer + 1), it will instead go to the last winner ((last winner -1) +1).  Needs to be tested.